table td

+++ src/api/list.rs
@@ -3,6 +3,6 @@ use crate::{error::Error, types::list::ListInfoWrapper};
 use super::request::request_get;
 
 /// Get list
-pub async fn all(page: u64) -> Result<ListInfoWrapper, Error> {
+pub async fn all(page: usize) -> Result<ListInfoWrapper, Error> {
     request_get::<ListInfoWrapper>(format!("/list?page={}", page)).await
 }
